\name{createBasicMNLFile}
\alias{createBasicMNLFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Creates a template control file for a basic multinomial logit (MNL) model with random coefficients
}
\description{
 Creates a template control file for a basic multinomial logit (MNL) model which a user can then customize to their needs. 
}
\usage{
 createBasicMNLFile()
}
\author{
 Jeffrey Dumont, RSG, Inc.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
