\name{plotA}
\alias{plotA}
\title{
 Plots the content of the A file
}
\description{
 This function reads and plots the means of the content of the A file. The A file contains the means of the underlying normal distributions for the random parameters.
}
\usage{
plotA(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{A string that corresponds to the model's name commonly stored in modelname. This the characters to the left of the _ in the A file name (model_A.csv)}
}
\author{
 Jeffrey Dumont, RSG, Inc.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
