\name{plotF}
\alias{plotF}
\title{
 Plots the content of the F file
}
\description{
 This function reads and plots the means of the content of the F file. The F file contains the fixed (non-random) parameters.
}
\usage{
plotF(model, columns=0)
}
\arguments{
  \item{model}{A string that corresponds to the model's name commonly stored in modelname. This the characters to the left of the _ in the F file name (model_F.csv)}
  \item{columns}{A vector of particular columns to plot. A default of 0 plots all columns.}
}
\author{
 Jeffrey Dumont, RSG, Inc.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
