\name{plotLog}
\alias{plotLog}
\title{
Plot Model Log Statistics
}
\description{
Creates a faceted plot of an \code{RSGHB} model object's iteration detail (statistics log).
}
\usage{
plotLog(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An \code{RSGHB} model object.}
}
\author{
 Jeffrey Dumont <jeff.dumont@rsginc.com>, Jeff Keller <jeff.keller@rsginc.com>
}
\seealso{
\code{\link{doHB}}, \code{\link{plotA}}, \code{\link{plotC}}, \code{\link{plotF}}
}
\examples{
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
