\name{RSGHB-internal}
\alias{checkModel}
\alias{hb}
\alias{nextA}
\alias{nextB}
\alias{nextD}
\alias{nextDind}
\alias{nextF}
\alias{prepareModel}
\alias{progreport}
\alias{trans}
\alias{vech}
\alias{xpnd}
\title{
 Internal RSGHB Functions
}
\description{
     Internal RSGHB functions
}
\details{
 These are not to be called by the user.
}
\author{
Jeffrey Dumont <jeff.dumont@rsginc.com>, Jeff Keller <jeff.keller>
}
\keyword{ internal }
