\name{vech}
\alias{vech}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Vectorizes a symmetric matrix
}
\description{
 Turns a symmetrix matrix into a matrix to save on memory space. The symmetric matrix is vectorizes by retaining the lower triangular portion of the the matrix.
}
\usage{
vech(mat)
}
\arguments{
  \item{mat}{ A symmetric matrix to be vectorized.}
}
\author{
 Jeffrey Dumont, RSG, Inc.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
