\name{xpnd}
\alias{xpnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Convert a vector into a symmetrix matrix
}
\description{
 Expands a vector into a symmetric matrix.
}
\usage{
xpnd(vec)
}
\arguments{
  \item{vec}{A vector to turn into a symmetric matrix}
}
\author{
Jeffrey Dumont, RSG, Inc.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
