\name{plotA}
\alias{plotA}
\title{
 Plot Population Means
}
\description{
This function plots the population means of the underlying normal distributions for the random parameters of an \code{RSGHB} model object.
}
\usage{
plotA(object, columns = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An \code{RSGHB} model object.}
  \item{columns}{A vector of particular columns to plot. If \code{NULL}, plots all columns. (Defaults to \code{NULL})}
}
\author{
 Jeffrey Dumont <jeff.dumont@rsginc.com>, Jeff Keller <jeff.keller@rsginc.com>
}
\seealso{
\code{\link{doHB}}, \code{\link{plotC}}, \code{\link{plotF}}, \code{\link{plotLog}}
}
\examples{
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
