\name{plotF}
\alias{plotF}
\title{
 Plot Fixed Parameters
}
\description{
Plots the fixed parameters in an \code{RSGHB} model object.
}
\usage{
plotF(object, columns = NULL)
}
\arguments{
  \item{object}{An \code{RSGHB} model object.}
  \item{columns}{A vector of particular columns to plot. If \code{NULL}, plots all columns. (Defaults to \code{NULL})}
}
\author{
 Jeffrey Dumont <jeff.dumont@rsginc.com>, Jeff Keller <jeff.keller@rsginc.com>
}
\seealso{
\code{\link{doHB}}, \code{\link{plotA}}, \code{\link{plotC}}, \code{\link{plotLog}}
}
\examples{
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
