\name{plotLog}
\alias{plotLog}
\title{
 Plots statistics contained in a model's the log file
}
\description{
 This function reads the log file and plots the log-likelihood, RLH and Parameter RMS for a given model.
}
\usage{
plotLog(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{A string that corresponds to the model's name commonly stored in modelname. This the characters to the left of .log in the file name (eg. model.log)}
}
\author{
 Jeffrey Dumont, RSG, Inc.
}
\note{

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
