\name{RSGHB-internal}
\alias{checkModel}
\alias{hb}
\alias{nextA}
\alias{nextB}
\alias{nextD}
\alias{nextDind}
\alias{nextF}
\alias{prepareModel}
\alias{trans}
\alias{writeLog}
\title{
 Internal RSGHB Functions
}
\description{
     Internal RSGHB functions
}
\details{
 These are not to be called by the user.
}
\author{
Jeffrey Dumont, RSG, Inc.
}
\keyword{ internal }
